requirements.txt
torch
transformers
datasets
scikit-learn
streamlit
python-whois
requests
fastapi
uvicorn
accelerate
python-dotenv
